sidelobe levels are correct
should have realized this earlier but since power can't interact with power this must be outputting relative e fields which means power is proportional to the square of the output of the simulator. If I do that the sidelobe levels are correct. There's still the issue of the non symmetric planar array output but that's probably a code issue not an algorithm issue